# Add prefetching to other recipe list view. Block DB access in serializers.