--- a/backend/core/recipes/views.py
+++ b/backend/core/recipes/views.py
@@ -139,7 +139,9 @@ class RecipeViewSet(viewsets.ModelViewSet):
         prefetched_recipe: Recipe = Recipe.objects.prefetch_related(
             "owner", "step_set", "ingredient_set", "scheduledrecipe_set"
         ).get(id=new_recipe.id)
-        return Response(RecipeSerializer(prefetched_recipe).data, status=status.HTTP_200_OK)
+        return Response(
+            RecipeSerializer(prefetched_recipe).data, status=status.HTTP_200_OK
+        )
 
 
 class TeamRecipesViewSet(