Disable DB access in serializers to prevent N+1 queries.

Teams recipes list view performance

name | trial 1 | trial 2 | trial 3 | avg | % change
---|---|---|---|--- |---
before | 0.286s | 0.180s | 0.279s | 0.248s| —
after | 0.169s | 0.156s | 0.167s | 0.164s | -51.2%